Fix dependencies expansion: handle active=True correctly
active=True in a module's manifest is a deprecated equivalent to auto_install=True:
https://github.com/odoo/odoo/commit/2cc7f0dd49a6be1920cef232e6c010fbafa68270#diff-367842087e68d6811cf53f2ab3eb21103cd8e793b1648ef3e57f220ee0ccde4eL345

it does not mean that Odoo installs it automatically, as mentioned in https://github.com/acsone/click-odoo-contrib/pull/9

Instead, we should handle that key the same way we handle the auto_install one.

Also, add a new simple test for this case.