Fix dependencies expansion: handle active and auto_install keywords correctly
According to the updated documentation, active=True in a module's manifest is a deprecated equivalent to auto_install=True:
https://github.com/odoo/odoo/commit/2cc7f0dd49a6be1920cef232e6c010fbafa68270#diff-367842087e68d6811cf53f2ab3eb21103cd8e793b1648ef3e57f220ee0ccde4eL345
https://www.odoo.com/documentation/15.0/developer/reference/backend/module.html

it does not mean that Odoo installs it automatically, as mentioned in https://github.com/acsone/click-odoo-contrib/pull/9

OTOH, currently Odoo ignores it. See:
https://github.com/odoo/odoo/blob/13.0/odoo/modules/module.py#L335
https://github.com/odoo/odoo/blob/13.0/odoo/modules/module.py#L370

So, we should ignore that key.

Besides, this implements the more complex functionality of defining auto_install as a subset of the module's dependencies, as also explained in the documentation.

Also, add a new simple test for this case.